// components/centraldashboard/app/clients/profile_controller.ts
export class Binding {
    'user'?: Subject;
    'referredNamespace'?: string;
    'roleRef'?: RoleRef;
    /**
    * Status of the profile, one of Succeeded, Failed, Unknown.
    */
    'status'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "user",
            "baseName": "user",
            "type": "Subject"
        },
        {
            "name": "referredNamespace",
            "baseName": "referredNamespace",
            "type": "string"
        },
        {
            "name": "roleRef",
            "baseName": "RoleRef",
            "type": "RoleRef"
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return Binding.attributeTypeMap;
    }
}
// ...
export class RoleRef {
    'apiGroup'?: string;
    'kind'?: string;
    'name'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "apiGroup",
            "baseName": "apiGroup",
            "type": "string"
        },
        {
            "name": "kind",
            "baseName": "kind",
            "type": "string"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return RoleRef.attributeTypeMap;
    }
}

/**
* in consistent with io.k8s.api.rbac.v1.Subject
*/
export class Subject {
    'kind'?: string;
    'name'?: string;
    'namespace'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "kind",
            "baseName": "kind",
            "type": "string"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "namespace",
            "baseName": "namespace",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return Subject.attributeTypeMap;
    }
}